feat(cart.js): add possibility to change product quantities directly in cart page
make it possible to change quantity of each product directly from cart page. When changing a
quantity, the total quantity of all articles and the total price are updated, as well as the cart
array and local storage. It is not yet possible to delete an item from cart neither by setting its
quantity at 0 nor by hitting delete button.